+++ src/neighbors/find_neighbors/run_test.py
@@ -21,7 +21,7 @@ logger.addHandler(console_handler)
 input = f"{meta['resources_dir']}/pbmc_1k_protein_v3/pbmc_1k_protein_v3_mms.h5mu"
 output = "output.h5mu"
 cmd_pars = [
-    "./" + meta["functionality_name"],
+    meta["executable"],
     "--input", input,
     "--output", output,
     "--obsp_name_prefix", "foo"